third_party/blink/web_tests/svg/dynamic-updates/SVGFEGaussianBlurElement-dom-stdDeviation-call-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' property of the SVGFEGaussianBlurElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS blurElement.stdDeviationX.baseVal is 5
PASS blurElement.stdDeviationY.baseVal is 5
PASS blurElement.stdDeviationX.baseVal is 10
PASS blurElement.stdDeviationY.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
